discord-bot-lambda/src/requirements.txt
flask
mangum
asgiref
discord-interactions
pyrebase4
python-dotenv